Move dialogs to more appropriate location
