+++ apps/frontend/src/app/5_pages/BorrowPage/BorrowPage.utils.ts
@@ -1,6 +1,7 @@
 import { SupportedTokens } from '@sovryn/contracts';
 
 import { BITCOIN } from '../../../constants/currencies';
+import { decimalic } from '../../../utils/math';
 
 export const isBtcBasedAsset = (asset: string) =>
   asset.toLowerCase() === SupportedTokens.rbtc ||
@@ -11,3 +12,16 @@ export const isBitpro = (asset: string) =>
   asset.toLowerCase() === 'bitpro' ||
   asset.toLowerCase() === 'bitp' ||
   asset.toLowerCase() === SupportedTokens.bpro;
+
+export const getCollateralAssetPrice = (
+  collateralToken: SupportedTokens,
+  debtToken: SupportedTokens,
+  rbtcPrice: string,
+  collateralPriceUsd: string,
+  borrowPriceUsd: string,
+) =>
+  decimalic(
+    collateralToken === SupportedTokens.rbtc ? rbtcPrice : collateralPriceUsd,
+  )
+    .div(debtToken === SupportedTokens.rbtc ? rbtcPrice : borrowPriceUsd)
+    .toString();
