+++ apps/frontend/src/app/5_pages/BorrowPage/components/AdjustLoanForm/hooks/useDepositCollateral.ts
@@ -6,6 +6,10 @@ import { SupportedTokens } from '@sovryn/contracts';
 
 import { defaultChainId } from '../../../../../../config/chains';
 
+import {
+  Transaction,
+  TransactionType,
+} from '../../../../../3_organisms/TransactionStepDialog/TransactionStepDialog.types';
 import { GAS_LIMIT } from '../../../../../../constants/gasLimits';
 import { useTransactionContext } from '../../../../../../contexts/TransactionContext';
 import { useAccount } from '../../../../../../hooks/useAccount';
@@ -13,37 +17,31 @@ import { useLoadContract } from '../../../../../../hooks/useLoadContract';
 import { translations } from '../../../../../../locales/i18n';
 import { toWei } from '../../../../../../utils/math';
 import { prepareApproveTransaction } from '../../../../../../utils/transactions';
-import {
-  Transaction,
-  TransactionType,
-} from '../../../../TransactionStepDialog/TransactionStepDialog.types';
 
-export const useRepayLoan = () => {
+export const useDepositCollateral = () => {
   const contract = useLoadContract('protocol', 'protocol', defaultChainId);
   const { setTransactions, setIsOpen, setTitle } = useTransactionContext();
   const { account, signer } = useAccount();
 
   const handleSubmit = useCallback(
     async (
-      repayAmount: string,
+      depositAmount: string,
+      depositToken: SupportedTokens,
       loanId: string,
-      borrowToken: SupportedTokens,
-      isPartialRepay?: boolean,
     ) => {
       if (!contract || !account || !signer) {
         return;
       }
 
-      const weiRepayAmount = toWei(repayAmount);
-
-      const isBorrowTokenRbtc = borrowToken === SupportedTokens.rbtc;
+      const weiDepositAmount = toWei(depositAmount);
+      const isDepositTokenRbtc = depositToken === SupportedTokens.rbtc;
 
       const transactions: Transaction[] = [];
 
-      if (!isBorrowTokenRbtc) {
+      if (!isDepositTokenRbtc) {
         const approve = await prepareApproveTransaction({
-          token: borrowToken,
-          amount: weiRepayAmount,
+          token: depositToken,
+          amount: weiDepositAmount,
           signer,
           spender: contract.address,
         });
@@ -54,16 +52,15 @@ export const useRepayLoan = () => {
 
       transactions.push({
         title: t(
-          translations.fixedInterestPage.adjustLoanDialog.dialogTitles[
-            isPartialRepay ? 'repay' : 'close'
-          ],
+          translations.fixedInterestPage.adjustLoanDialog.dialogTitles
+            .addCollateral,
         ),
         request: {
           type: TransactionType.signTransaction,
           contract,
-          fnName: 'closeWithDeposit',
-          args: [loanId, account, weiRepayAmount],
-          value: isBorrowTokenRbtc ? weiRepayAmount : '0',
+          fnName: 'depositCollateral',
+          args: [loanId, weiDepositAmount],
+          value: isDepositTokenRbtc ? weiDepositAmount : '0',
           gasLimit: GAS_LIMIT.REPAY_LOAN,
         },
       });
@@ -71,14 +68,13 @@ export const useRepayLoan = () => {
       setTransactions(transactions);
       setTitle(
         t(
-          translations.fixedInterestPage.adjustLoanDialog.dialogTitles[
-            isPartialRepay ? 'repay' : 'close'
-          ],
+          translations.fixedInterestPage.adjustLoanDialog.dialogTitles
+            .addCollateral,
         ),
       );
-      setTransactions(transactions);
       setIsOpen(true);
     },
+
     [account, contract, setIsOpen, setTitle, setTransactions, signer],
   );
