+++ apps/frontend/src/app/5_pages/BorrowPage/components/AdjustLoanForm/hooks/useGetBorrowingAPR.ts
