+++ apps/frontend/src/app/5_pages/BorrowPage/components/AdjustLoanForm/hooks/useWithdrawCollateral.ts
@@ -4,16 +4,16 @@ import { t } from 'i18next';
 
 import { defaultChainId } from '../../../../../../config/chains';
 
+import {
+  Transaction,
+  TransactionType,
+} from '../../../../../3_organisms/TransactionStepDialog/TransactionStepDialog.types';
 import { GAS_LIMIT } from '../../../../../../constants/gasLimits';
 import { useTransactionContext } from '../../../../../../contexts/TransactionContext';
 import { useAccount } from '../../../../../../hooks/useAccount';
 import { useLoadContract } from '../../../../../../hooks/useLoadContract';
 import { translations } from '../../../../../../locales/i18n';
 import { toWei } from '../../../../../../utils/math';
-import {
-  Transaction,
-  TransactionType,
-} from '../../../../TransactionStepDialog/TransactionStepDialog.types';
 
 export const useWithdrawCollateral = () => {
   const contract = useLoadContract('protocol', 'protocol', defaultChainId);
