+++ apps/frontend/src/app/5_pages/BorrowPage/components/AdvancedSettings/AdvancedSettings.tsx
