+++ apps/frontend/src/app/5_pages/BorrowPage/components/BorrowAssetsTable/BorrowAssetsTable.constants.tsx
@@ -5,11 +5,11 @@ import { t } from 'i18next';
 import { HelperButton } from '@sovryn/ui';
 
 import { AssetRenderer } from '../../../../2_molecules/AssetRenderer/AssetRenderer';
-import { normalizeToken } from '../../../../3_organisms/BorrowLoan/components/AdjustLoanForm/AdjustLoanForm.utils';
 import { MINIMUM_COLLATERAL_RATIO_BORROWING_MAINTENANCE } from '../../../../../constants/lending';
 import { translations } from '../../../../../locales/i18n';
 import { LendingPool } from '../../../../../utils/LendingPool';
 import { AcceptedCollateral } from '../../../LendPage/components/AcceptedCollateral/AcceptedCollateral';
+import { normalizeToken } from '../AdjustLoanForm/AdjustLoanForm.utils';
 import { AvailableSupply } from './components/AvailableSupply/AvailableSupply';
 import { NewLoanButton } from './components/NewLoanButton/NewLoanButton';
 import { NextBorrowInterestRate } from './components/NextBorrowInterestRate/NextBorrowInterestRate';
