+++ apps/frontend/src/app/5_pages/BorrowPage/components/BorrowAssetsTable/components/NextBorrowInterestRate/NextBorrowInterestRate.tsx
@@ -4,7 +4,7 @@ import { SupportedTokens } from '@sovryn/contracts';
 import { Decimal } from '@sovryn/utils';
 
 import { AmountRenderer } from '../../../../../../2_molecules/AmountRenderer/AmountRenderer';
-import { useGetBorrowingAPR } from '../../../../../../3_organisms/BorrowLoan/components/AdjustLoanForm/hooks/useGetBorrowingAPR';
+import { useGetBorrowingAPR } from '../../../AdjustLoanForm/hooks/useGetBorrowingAPR';
 
 type NextBorrowInterestRateProps = {
   asset: SupportedTokens;
