+++ apps/frontend/src/app/5_pages/BorrowPage/components/ExtendLoanForm/ExtendLoanForm.utils.tsx
