+++ apps/frontend/src/app/5_pages/BorrowPage/components/MaxDuration/MaxDuration.tsx
