+++ apps/frontend/src/app/5_pages/BorrowPage/components/NewLoanForm/NewLoanForm.utils.tsx
